Refactored construction of test corpus into utility class
